Refactored components to remove duplicate code.
